[resotocore][chore] Refactor aio.sleep in graph_merge test
